refactor(templates): split WorkflowTemplateBrowserModal into portable UI and Lightning wrapper
The original component mixed UI rendering with Phoenix channel fetching,
Y.Doc workflow imports, and global store hooks — making it impossible to
render or iterate on in isolation (Storybook, tests, or a plain React context)
without standing up the full collaborative editor infrastructure.

Split into two components:
- TemplateBrowserModal: pure presentational modal, takes only plain props,
  no Lightning dependencies. Product can open this file and iterate freely.
- TemplateBrowserModalWrapper: thin wiring layer that connects the modal to
  the Phoenix channel, Y.Doc workflow store, and global UI state.

Also fixes isSaving not being reset on successful template selection, and
clears the search query each time the modal opens.
